+++ hazelcast-sql/src/main/java/com/hazelcast/jet/sql/impl/LogicalTableInsert.java
@@ -28,7 +28,7 @@ import java.util.List;
 
 public class LogicalTableInsert extends TableModify {
 
-    LogicalTableInsert(TableModify modify) {
+    public LogicalTableInsert(TableModify modify) {
         this(
                 modify.getCluster(),
                 modify.getTraitSet(),
