--- hazelcast-sql/src/main/java/com/hazelcast/sql/impl/calcite/validate/HazelcastJetSqlConformance.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2021 Hazelcast Inc.
- *
- * Licensed under the Hazelcast Community License (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://hazelcast.com/hazelcast-community-license
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.hazelcast.sql.impl.calcite.validate;
-
-public final class HazelcastJetSqlConformance extends HazelcastSqlConformance {
-    public static final HazelcastJetSqlConformance INSTANCE = new HazelcastJetSqlConformance();
-
-    private HazelcastJetSqlConformance() {
-        super();
-    }
-
-    @Override
-    public boolean isFromRequired() {
-        return false;
-    }
-}
